requirements.txt
fastapi
uvicorn
requests
python-dotenv
urllib3<2.0
beautifulsoup4
